# Conflicts:
#	api/src/main/scala/hr/com/blanka/apartments/command/price/commandProtocol.scala
#	api/src/main/scala/hr/com/blanka/apartments/query/price/DailyPriceAggregateActor.scala


Former-commit-id: 77bc6efcb15300c337536cdc9a556033d3ced803 [formerly eee741921da90a5d86f467d723f2a544c4240431]
Former-commit-id: cca332bb4626b3a4092d828fb4d11c5f8a2810d8

+++ api/src/main/scala/hr/com/blanka/apartments/query/price/QueryPriceRangeActor.scala
@@ -19,7 +19,7 @@ object QueryPriceRangeActor {
   def apply(dailyPriceActor: ActorRef) = Props(classOf[QueryPriceRangeActor], dailyPriceActor)
 }
 
-case class CalculationData(singleDayCalculations: Map[Long, Option[Double]])
+case class CalculationData(singleDayCalculations: Map[Long, Option[BigDecimal]])
 
 class QueryPriceRangeActor(dailyPriceActor: ActorRef) extends Actor {
 
@@ -47,7 +47,7 @@ class QueryPriceRangeActor(dailyPriceActor: ActorRef) extends Actor {
       val newlySentDailyCalculationMessages = sendMessagesForSingleDayCalculations(cpfr)
 
       Future.sequence(newlySentDailyCalculationMessages).onComplete {
-        case Success(result) => msgSender ! Good(result.foldLeft(0.0)((sum, next) => next.asInstanceOf[PriceDayFetched].price + sum))
+        case Success(result) => msgSender ! Good(result.foldLeft(BigDecimal(0))((sum, next) => next.asInstanceOf[PriceDayFetched].price + sum))
         case Failure(t) => msgSender ! Bad("An error has occurred: " + t.getMessage)
       }
   }
